automations/projeto_ad/requirements.txt
cryptography
mysql-connector-python
pymssql
PyMySQL
configparser
zeep
openpyxl
pandas

